Implemented an asset handling trait
Refactoring to increase the utiliaztion of traits and
interfaces to streamline bacon as part of the 2.0 reoadmap.
