+++ poc/test-asset.php
@@ -1,16 +1,16 @@
 <?php
 
-class TestAsset extends BtfJsAsset {
-    const ASSET_ID      = 'test-asset';
-    const ASSET_DATA_ID = 'test_data';
+class Test_Asset extends Btf_Js_Asset {
+	const ASSET_ID      = 'test-asset';
+	const ASSET_DATA_ID = 'test_data';
 
-    public $data = array(
-            "lotame_id" => "LOTCC_10234"
-        );
+	public $data = array(
+			"lotame_id" => "LOTCC_10234"
+		);
 
-    public function __construct() {
-        $this->register_local_data( $this->data );
-    }
+	public function __construct() {
+		$this->register_local_data( $this->data );
+	}
 
 }
